handlers.py, utils.py.enterContinue() update
