+++ Library/__init__.py
@@ -1,4 +1,5 @@
 from .utils import *
 from .jsonHelper import *
 from .entities import *
+from .handlers import *
 from .game import *
