+++ Library/game.py
@@ -1,4 +1,4 @@
-from Library import utils, entities
+from Library import utils, entities, handlers
 
 # Labels
 # TODO use these vars to provide different versions of the messages later, using data saved in something like lines.json
@@ -17,7 +17,7 @@ def play():
     player.hp = 10 # TODO difficulty options, affect damage as well, affect also the enemy, both hp and damage
     utils.intro()
     player.name = input(playerNameMessage)
-    utils.enterContinue(dangerousJourneyMessage, False)
+    utils.enterContinue(dangerousJourneyMessage, False, True)
     loop(player)
 
 def loop(player):    
@@ -27,10 +27,10 @@ def loop(player):
     print(creatureSpawned%(creature))
     if creature.hp == 0:
         # TODO custom death messages, make this very rare to happen as well
-        utils.enterContinue(tooWeak%creature,False)
+        utils.enterContinue(tooWeak%creature, False, True)
         loop(player)
     else:
-        utils.handlePlayerInput(player, creature)
+        handlers.handlePlayerInput(player, creature)
     handleResult(player, creature)
     
 def handleResult(player, creature):
@@ -41,6 +41,6 @@ def handleResult(player, creature):
     if action == "y":
         loop(player)
     elif action == "n":
-        utils.handleVictory(player)
+        handlers.handleVictory(player)
     else:
         handleResult(player, creature)
